Rename the `src/display/dom_utils.js` file to `src/display/display_utils.js`
This file (currently) contains not only DOM-specific helper functions/classes, but is used generally for various helper code relevant for main-thread functionality.
