+++ test/unit/clitests.json
@@ -14,8 +14,8 @@
     "colorspace_spec.js",
     "crypto_spec.js",
     "display_svg_spec.js",
+    "display_utils_spec.js",
     "document_spec.js",
-    "dom_utils_spec.js",
     "encodings_spec.js",
     "evaluator_spec.js",
     "function_spec.js",
