+++ test/unit/jasmine-boot.js
@@ -56,8 +56,8 @@ function initializePDFJS(callback) {
     'pdfjs-test/unit/crypto_spec',
     'pdfjs-test/unit/custom_spec',
     'pdfjs-test/unit/display_svg_spec',
+    'pdfjs-test/unit/display_utils_spec',
     'pdfjs-test/unit/document_spec',
-    'pdfjs-test/unit/dom_utils_spec',
     'pdfjs-test/unit/encodings_spec',
     'pdfjs-test/unit/evaluator_spec',
     'pdfjs-test/unit/function_spec',
